Move metric server installation to container
This adds and entrypoint to docker image that checks if metrics server is already installed.

If not installed it will install metrics server on devzero-zxporter namespace.

This also adds the required permissions to isntall metrics server. This means we now need some writing permissions but they are mostly restricted to devzero namespace. Some permissions are cluster wide and are needed to install metrics server.

We could potentially try to trim permissions in the future.